cmd: Move ./gotpm to ./cmd/gotpm
Our current filesystem structure is non-standard. Normally, go modules
have a "cmd" directory containing all of the binaries built as
subdirectories. We invert this.

Fix to be more standard.
