+++ cmd/flush.go
