+++ cmd/flush_test.go
