+++ cmd/pubkey.go
